sink(ticdc): refactor canal and canal-json protocols and support callback
