+++ cdc/sink/mq/codec/canal_json_encoder_test.go
@@ -1,0 +1,301 @@
+// Copyright 2022 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package codec
+
+import (
+	"context"
+	"encoding/json"
+	"testing"
+
+	"github.com/pingcap/tiflow/cdc/model"
+	"github.com/pingcap/tiflow/pkg/config"
+	"github.com/stretchr/testify/require"
+	"golang.org/x/text/encoding/charmap"
+)
+
+func TestBuildCanalJSONBatchEncoder(t *testing.T) {
+	t.Parallel()
+	cfg := NewConfig(config.ProtocolCanalJSON)
+
+	builder := &canalJSONBatchEncoderBuilder{config: cfg}
+	encoder, ok := builder.Build().(*canalJSONBatchEncoder)
+	require.True(t, ok)
+	require.False(t, encoder.enableTiDBExtension)
+
+	cfg.enableTiDBExtension = true
+	builder = &canalJSONBatchEncoderBuilder{config: cfg}
+	encoder, ok = builder.Build().(*canalJSONBatchEncoder)
+	require.True(t, ok)
+	require.True(t, encoder.enableTiDBExtension)
+}
+
+func TestNewCanalJSONMessage4DML(t *testing.T) {
+	t.Parallel()
+	encoder := &canalJSONBatchEncoder{builder: newCanalEntryBuilder()}
+	require.NotNil(t, encoder)
+
+	message, err := encoder.newJSONMessageForDML(testCaseInsert)
+	require.Nil(t, err)
+	jsonMsg, ok := message.(*canalJSONMessage)
+	require.True(t, ok)
+	require.NotNil(t, jsonMsg.Data)
+	require.Nil(t, jsonMsg.Old)
+	require.Equal(t, "INSERT", jsonMsg.EventType)
+	require.Equal(t, convertToCanalTs(testCaseInsert.CommitTs), jsonMsg.ExecutionTime)
+	require.Equal(t, testCaseInsert.CommitTs, jsonMsg.tikvTs)
+	require.Equal(t, "cdc", jsonMsg.Schema)
+	require.Equal(t, "person", jsonMsg.Table)
+	require.False(t, jsonMsg.IsDDL)
+
+	// check data is enough
+	obtainedDataMap := jsonMsg.getData()
+	require.NotNil(t, obtainedDataMap)
+
+	for _, item := range testColumnsTable {
+		obtainedValue, ok := obtainedDataMap[item.column.Name]
+		require.True(t, ok)
+		if !item.column.Flag.IsBinary() {
+			require.Equal(t, item.expectedEncodedValue, obtainedValue)
+			continue
+		}
+
+		// for `Column.Value` is nil, which mean's it is nullable, set the value to `""`
+		if obtainedValue == nil {
+			require.Equal(t, "", item.expectedEncodedValue)
+			continue
+		}
+
+		if bytes, ok := item.column.Value.([]byte); ok {
+			expectedValue, err := charmap.ISO8859_1.NewDecoder().Bytes(bytes)
+			require.Nil(t, err)
+			require.Equal(t, string(expectedValue), obtainedValue)
+			continue
+		}
+
+		require.Equal(t, item.expectedEncodedValue, obtainedValue)
+	}
+
+	message, err = encoder.newJSONMessageForDML(testCaseUpdate)
+	require.Nil(t, err)
+	jsonMsg, ok = message.(*canalJSONMessage)
+	require.True(t, ok)
+	require.NotNil(t, jsonMsg.Data)
+	require.NotNil(t, jsonMsg.Old)
+	require.Equal(t, "UPDATE", jsonMsg.EventType)
+
+	message, err = encoder.newJSONMessageForDML(testCaseDelete)
+	require.Nil(t, err)
+	jsonMsg, ok = message.(*canalJSONMessage)
+	require.True(t, ok)
+	require.NotNil(t, jsonMsg.Data)
+	require.Nil(t, jsonMsg.Old)
+	require.Equal(t, "DELETE", jsonMsg.EventType)
+
+	encoder = &canalJSONBatchEncoder{builder: newCanalEntryBuilder(), enableTiDBExtension: true}
+	require.NotNil(t, encoder)
+	message, err = encoder.newJSONMessageForDML(testCaseUpdate)
+	require.Nil(t, err)
+
+	withExtension, ok := message.(*canalJSONMessageWithTiDBExtension)
+	require.True(t, ok)
+
+	require.NotNil(t, withExtension.Extensions)
+	require.Equal(t, testCaseUpdate.CommitTs, withExtension.Extensions.CommitTs)
+}
+
+func TestNewCanalJSONMessageFromDDL(t *testing.T) {
+	t.Parallel()
+	encoder := &canalJSONBatchEncoder{builder: newCanalEntryBuilder()}
+	require.NotNil(t, encoder)
+
+	message := encoder.newJSONMessageForDDL(testCaseDDL)
+	require.NotNil(t, message)
+
+	msg, ok := message.(*canalJSONMessage)
+	require.True(t, ok)
+	require.Equal(t, testCaseDDL.CommitTs, msg.tikvTs)
+	require.Equal(t, convertToCanalTs(testCaseDDL.CommitTs), msg.ExecutionTime)
+	require.True(t, msg.IsDDL)
+	require.Equal(t, "cdc", msg.Schema)
+	require.Equal(t, "person", msg.Table)
+	require.Equal(t, testCaseDDL.Query, msg.Query)
+	require.Equal(t, "CREATE", msg.EventType)
+
+	encoder = &canalJSONBatchEncoder{builder: newCanalEntryBuilder(), enableTiDBExtension: true}
+	require.NotNil(t, encoder)
+
+	message = encoder.newJSONMessageForDDL(testCaseDDL)
+	require.NotNil(t, message)
+
+	withExtension, ok := message.(*canalJSONMessageWithTiDBExtension)
+	require.True(t, ok)
+
+	require.NotNil(t, withExtension.Extensions)
+	require.Equal(t, testCaseDDL.CommitTs, withExtension.Extensions.CommitTs)
+}
+
+func TestBatching(t *testing.T) {
+	t.Parallel()
+	encoder := &canalJSONBatchEncoder{builder: newCanalEntryBuilder()}
+	require.NotNil(t, encoder)
+
+	updateCase := *testCaseUpdate
+	for i := 1; i <= 1000; i++ {
+		ts := uint64(i)
+		updateCase.CommitTs = ts
+		err := encoder.AppendRowChangedEvent(context.Background(), "", &updateCase, nil)
+		require.Nil(t, err)
+
+		if i%100 == 0 {
+			msgs := encoder.Build()
+			require.NotNil(t, msgs)
+			require.Len(t, msgs, 100)
+
+			for j := range msgs {
+				require.Equal(t, 1, msgs[j].GetRowsCount())
+
+				var msg canalJSONMessage
+				err := json.Unmarshal(msgs[j].Value, &msg)
+				require.Nil(t, err)
+				require.Equal(t, "UPDATE", msg.EventType)
+			}
+		}
+	}
+
+	require.Len(t, encoder.messageBuf, 0)
+}
+
+func TestEncodeCheckpointEvent(t *testing.T) {
+	t.Parallel()
+	var watermark uint64 = 2333
+	for _, enable := range []bool{false, true} {
+		encoder := &canalJSONBatchEncoder{builder: newCanalEntryBuilder(), enableTiDBExtension: enable}
+		require.NotNil(t, encoder)
+
+		msg, err := encoder.EncodeCheckpointEvent(watermark)
+		require.Nil(t, err)
+
+		if !enable {
+			require.Nil(t, msg)
+			continue
+		}
+
+		require.NotNil(t, msg)
+		decoder := NewCanalJSONBatchDecoder(msg.Value, enable)
+
+		ty, hasNext, err := decoder.HasNext()
+		require.Nil(t, err)
+		if enable {
+			require.True(t, hasNext)
+			require.Equal(t, model.MessageTypeResolved, ty)
+			consumed, err := decoder.NextResolvedEvent()
+			require.Nil(t, err)
+			require.Equal(t, watermark, consumed)
+		} else {
+			require.False(t, hasNext)
+			require.Equal(t, model.MessageTypeUnknown, ty)
+		}
+
+		ty, hasNext, err = decoder.HasNext()
+		require.Nil(t, err)
+		require.False(t, hasNext)
+		require.Equal(t, model.MessageTypeUnknown, ty)
+	}
+}
+
+func TestCheckpointEventValueMarshal(t *testing.T) {
+	t.Parallel()
+	var watermark uint64 = 1024
+	encoder := &canalJSONBatchEncoder{
+		builder:             newCanalEntryBuilder(),
+		enableTiDBExtension: true,
+	}
+	require.NotNil(t, encoder)
+	msg, err := encoder.EncodeCheckpointEvent(watermark)
+	require.Nil(t, err)
+	require.NotNil(t, msg)
+
+	// Unmarshal from the data we have encoded.
+	jsonMsg := canalJSONMessageWithTiDBExtension{
+		&canalJSONMessage{},
+		&tidbExtension{},
+	}
+	err = json.Unmarshal(msg.Value, &jsonMsg)
+	require.Nil(t, err)
+	require.Equal(t, watermark, jsonMsg.Extensions.WatermarkTs)
+	// Hack the build time.
+	// Otherwise, the timing will be inconsistent.
+	jsonMsg.BuildTime = 1469579899
+	rawBytes, err := json.MarshalIndent(jsonMsg, "", "  ")
+	require.Nil(t, err)
+
+	// No commit ts will be output.
+	expectedJSON := `{
+  "id": 0,
+  "database": "",
+  "table": "",
+  "pkNames": null,
+  "isDdl": false,
+  "type": "TIDB_WATERMARK",
+  "es": 0,
+  "ts": 1469579899,
+  "sql": "",
+  "sqlType": null,
+  "mysqlType": null,
+  "data": null,
+  "old": null,
+  "_tidb": {
+    "watermarkTs": 1024
+  }
+}`
+	require.Equal(t, expectedJSON, string(rawBytes))
+}
+
+func TestDDLEventWithExtensionValueMarshal(t *testing.T) {
+	t.Parallel()
+	encoder := &canalJSONBatchEncoder{builder: newCanalEntryBuilder(), enableTiDBExtension: true}
+	require.NotNil(t, encoder)
+
+	message := encoder.newJSONMessageForDDL(testCaseDDL)
+	require.NotNil(t, message)
+
+	msg, ok := message.(*canalJSONMessageWithTiDBExtension)
+	require.True(t, ok)
+	// Hack the build time.
+	// Otherwise, the timing will be inconsistent.
+	msg.BuildTime = 1469579899
+	rawBytes, err := json.MarshalIndent(msg, "", "  ")
+	require.Nil(t, err)
+
+	// No watermark ts will be output.
+	expectedJSON := `{
+  "id": 0,
+  "database": "cdc",
+  "table": "person",
+  "pkNames": null,
+  "isDdl": true,
+  "type": "CREATE",
+  "es": 1591943372224,
+  "ts": 1469579899,
+  "sql": "create table person(id int, name varchar(32), tiny tinyint unsigned, comment text, primary key(id))",
+  "sqlType": null,
+  "mysqlType": null,
+  "data": null,
+  "old": null,
+  "_tidb": {
+    "commitTs": 417318403368288260
+  }
+}`
+	require.Equal(t, expectedJSON, string(rawBytes))
+}
