+++ cdc/sink/mq/codec/encoder.go
@@ -51,13 +51,13 @@ func NewEventBatchEncoderBuilder(ctx context.Context, c *Config) (EncoderBuilder
 	case config.ProtocolDefault, config.ProtocolOpen:
 		return newOpenProtocolBatchEncoderBuilder(c), nil
 	case config.ProtocolCanal:
-		return newCanalEventBatchEncoderBuilder(), nil
+		return newCanalBatchEncoderBuilder(), nil
 	case config.ProtocolAvro:
 		return newAvroEventBatchEncoderBuilder(ctx, c)
 	case config.ProtocolMaxwell:
 		return newMaxwellEventBatchEncoderBuilder(), nil
 	case config.ProtocolCanalJSON:
-		return newCanalFlatEventBatchEncoderBuilder(c), nil
+		return newCanalJSONBatchEncoderBuilder(c), nil
 	case config.ProtocolCraft:
 		return newCraftEventBatchEncoderBuilder(c), nil
 	default:
